+++ src/publish-packages-e2e.ts
@@ -1,5 +1,6 @@
 import path from 'path';
 import { publish } from 'rocket-punch';
+import { readEntry } from 'rocket-punch/entry/readEntry';
 import { publishMessageHandler } from 'rocket-punch/message-handlers/publish';
 import yargs from 'yargs';
 
@@ -7,7 +8,8 @@ const argv = yargs.argv;
 
 publish({
   cwd: process.cwd(),
-  dist: path.join(process.cwd(), 'dist'),
+  entry: readEntry({ cwd: process.cwd() }),
+  dist: path.join(process.cwd(), 'out/packages'),
   skipSelection: true,
   tag: argv.tag as string,
   registry: argv.registry as string,