+++ src/rocket-punch/__tests__/getPackagesEntry.test.ts
@@ -1,7 +1,8 @@
 import { copyTmpDirectory } from '@ssen/tmp-directory';
 import process from 'process';
 import { PackageInfo } from 'rocket-punch';
-import { getPackagesEntry } from 'rocket-punch/entry/getPackagesEntry';
+import { readEntry } from 'rocket-punch/entry/readEntry';
+import { readPackages } from 'rocket-punch/entry/readPackages';
 import { getRootDependencies } from 'rocket-punch/package-json/getRootDependencies';
 import { PackageJson } from 'type-fest';
 
@@ -11,17 +12,17 @@ describe('getPackagesEntry()', () => {
     const cwd: string = await copyTmpDirectory(process.cwd(), `test/fixtures/rocket-punch/sample`);
 
     // Act
-    const entry: Map<string, PackageInfo> = await getPackagesEntry({ cwd });
+    const packages: Map<string, PackageInfo> = await readPackages({ cwd, entry: readEntry({ cwd }) });
     const externalPackages: PackageJson.Dependency = await getRootDependencies({ cwd });
 
     // Assert
-    expect(entry.get('a')?.module).toBe('commonjs');
-    expect(entry.get('a')?.packageJson).toEqual({});
-    expect(entry.get('a')?.compilerOptions).toEqual({});
+    expect(packages.get('a')?.module).toBe('commonjs');
+    expect(packages.get('a')?.packageJson).toEqual({});
+    expect(packages.get('a')?.compilerOptions).toEqual({});
 
-    expect(entry.has('a')).toBeTruthy();
-    expect(entry.has('b')).toBeTruthy();
-    expect(entry.has('c')).toBeTruthy();
+    expect(packages.has('a')).toBeTruthy();
+    expect(packages.has('b')).toBeTruthy();
+    expect(packages.has('c')).toBeTruthy();
     expect('react' in externalPackages).toBeTruthy();
   });
 
@@ -30,17 +31,17 @@ describe('getPackagesEntry()', () => {
     const cwd: string = await copyTmpDirectory(process.cwd(), `test/fixtures/rocket-punch/group-entry`);
 
     // Act
-    const entry: Map<string, PackageInfo> = await getPackagesEntry({ cwd });
+    const packages: Map<string, PackageInfo> = await readPackages({ cwd, entry: readEntry({ cwd }) });
     const externalPackages: PackageJson.Dependency = await getRootDependencies({ cwd });
 
     // Assert
-    expect(entry.get('@group/a')?.module).toBe('commonjs');
-    expect(entry.get('@group/a')?.packageJson).toEqual({});
-    expect(entry.get('@group/a')?.compilerOptions).toEqual({});
+    expect(packages.get('@group/a')?.module).toBe('commonjs');
+    expect(packages.get('@group/a')?.packageJson).toEqual({});
+    expect(packages.get('@group/a')?.compilerOptions).toEqual({});
 
-    expect(entry.has('@group/a')).toBeTruthy();
-    expect(entry.has('@group/b')).toBeTruthy();
-    expect(entry.has('@group/c')).toBeTruthy();
+    expect(packages.has('@group/a')).toBeTruthy();
+    expect(packages.has('@group/b')).toBeTruthy();
+    expect(packages.has('@group/c')).toBeTruthy();
     expect('react' in externalPackages).toBeTruthy();
   });
 
@@ -52,19 +53,19 @@ describe('getPackagesEntry()', () => {
     );
 
     // Act
-    const entry: Map<string, PackageInfo> = await getPackagesEntry({ cwd });
+    const packages: Map<string, PackageInfo> = await readPackages({ cwd, entry: readEntry({ cwd }) });
 
     // Assert
-    expect(entry.get('a')?.module).toBe('commonjs');
-    expect(entry.get('a')?.packageJson['bin']['cli-a']).toBe('./bin/test.js');
-    expect(entry.get('b')?.module).toBe('commonjs');
-    expect(entry.get('b')?.packageJson['test']).toBe('hello world 1');
-    expect(entry.get('b')?.compilerOptions['allowJs']).toBeTruthy();
-    expect(entry.get('c')?.module).toBe('commonjs');
-    expect(entry.get('c')?.packageJson['test']).toBe('hello world 2');
+    expect(packages.get('a')?.module).toBe('commonjs');
+    expect(packages.get('a')?.packageJson['bin']['cli-a']).toBe('./bin/test.js');
+    expect(packages.get('b')?.module).toBe('commonjs');
+    expect(packages.get('b')?.packageJson['test']).toBe('hello world 1');
+    expect(packages.get('b')?.compilerOptions['allowJs']).toBeTruthy();
+    expect(packages.get('c')?.module).toBe('commonjs');
+    expect(packages.get('c')?.packageJson['test']).toBe('hello world 2');
 
-    expect(entry.has('a')).toBeTruthy();
-    expect(entry.has('b')).toBeTruthy();
-    expect(entry.has('c')).toBeTruthy();
+    expect(packages.has('a')).toBeTruthy();
+    expect(packages.has('b')).toBeTruthy();
+    expect(packages.has('c')).toBeTruthy();
   });
 });