+++ src/rocket-punch/rule/fileNames.ts
@@ -1,8 +1,2 @@
 export const packagesFileName: string = '.packages.yaml';
 export const sharedPackageJsonFileName: string = '.package.json';
-
-export const packageTransformFile: string = '.package';
-export const packageTransformFilePattern: RegExp = /.package.(js|ts)$/;
-
-//export const packageJsonFactoryFileName: string = '.package.json';
-//export const packageJsonFactoryFileNamePattern: RegExp = /.package.json.(js|ts)$/;