+++ test/fixtures/rocket-punch/transform-package-json/src/b/.package.ts
@@ -1,8 +1,0 @@
-import { PackageJsonTransformFunction } from 'rocket-punch';
-
-export const transformPackageJson: PackageJsonTransformFunction = (computedPackageJson) => {
-  return {
-    ...computedPackageJson,
-    keywords: ['hello'],
-  };
-};