+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Patched...
+## [2.0.0] - 2020-08-11
+
+### Breaking Changes
+- Change default output directory from `/dist` to `/out/packages`
+- ⚠️ Remove `.package.ts` rule
+- Add `transformPackageJson`, `transformCompilerOptions`, `transformCompilerHost` and `emitCustomTransformers` options to `build()`. these options replace the old `.package.ts` file.
 
 ### Added
 - Add an option `rocket-punch build --svg default`
@@ -64,6 +69,7 @@ Release
 ### Added
 - Rewrite paths by TypeScript compiler transformer API on build (e.g. `import A from '../a'` to `import A from 'a'`)
 
+[2.0.0]: https://github.com/rocket-hangar/rocket-punch/compare/v1.2.1...v2.0.0
 [1.3.1]: https://github.com/rocket-hangar/rocket-punch/compare/v1.1.0...v1.3.1
 [1.1.0]: https://github.com/rocket-hangar/rocket-punch/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/rocket-hangar/rocket-punch/compare/v1.0.0-alpha.9...v1.0.0